fix: move Itinerary sub components to own files, remove dead props
